+++ m2c.scm.domain/src/main/java/cn/m2c/scm/domain/model/order/SimpleMarketInfo.java
@@ -14,7 +14,7 @@ public class SimpleMarketInfo extends ValueObject {
 	/**营销层级*/
 	private Integer marketLevel;
 	/**营销层级门槛*/
-	private Integer threshold;
+	private long threshold;
 	/**营销形式，1：减钱，2：打折，3：换购*/
 	private Integer marketType;
 	/**门槛类型，1：金额，2：件数*/
@@ -30,7 +30,7 @@ public class SimpleMarketInfo extends ValueObject {
 		super();
 	}
 	
-	public SimpleMarketInfo(String marketingId, int marketLevel, int threshold,
+	public SimpleMarketInfo(String marketingId, int marketLevel, long threshold,
 			int marketType, int thresholdType, String sharePercent, String marketName
 			, int discount) {
 		super();
